+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### Unreleased
+### [v0.106](https://github.com/realityforge/guiceyloops/tree/v0.106) (2020-01-17) · [Full Changelog](https://github.com/realityforge/guiceyloops/compare/v0.105...v0.106)
 
 * Add `ValueUtil.addSeconds(...)` utility method.
 * Add `ValueUtil.minus...()` utility methods for every `ValueUtil.add...()` method.